20200725 1、手动封装mongo泛型service 2、表单定义/表单类型 MO定义 3、数据库int 转tinyint

+++ egg-common/src/main/java/com/egg/manager/common/base/enums/base/BaseStateEnum.java
@@ -9,27 +9,27 @@ package com.egg.manager.common.base.enums.base;
  * \
  */
 public enum  BaseStateEnum {
-    ENABLED(1,"启用","启用"),
-    DISABLED(0,"禁用","禁用"),
-    DELETE(-1,"删除","删除"),
-    LOCKED(-10,"锁定","锁定")
+    ENABLED((short)1,"启用","启用"),
+    DISABLED((short)0,"禁用","禁用"),
+    DELETE((short)-1,"删除","删除"),
+    LOCKED((short)-10,"锁定","锁定")
     ;
 
-    BaseStateEnum(Integer value, String name, String info) {
+    BaseStateEnum(Short value, String name, String info) {
         this.value = value;
         this.name = name;
         this.info = info;
     }
 
-    private Integer value ;
+    private Short value ;
     private String name ;
     private String info ;
 
-    public Integer getValue() {
+    public Short getValue() {
         return value;
     }
 
-    public void setValue(Integer value) {
+    public void setValue(Short value) {
         this.value = value;
     }
