+++ egg-common/src/main/java/com/egg/manager/common/base/enums/user/UserAccountStateEnum.java
@@ -1,24 +1,24 @@
 package com.egg.manager.common.base.enums.user;
 
 public enum UserAccountStateEnum {
-    ENABLED(1,"启用","启用"),
-    DISABLED(0,"禁用","禁用"),
-    DELETE(-1,"删除","用户已被删除"),
-    LOCKED(-10,"锁定","用户已被锁定"),
+    ENABLED((short) 1,"启用","启用"),
+    DISABLED((short)0,"禁用","禁用"),
+    DELETE((short)-1,"删除","用户已被删除"),
+    LOCKED((short)-10,"锁定","用户已被锁定"),
     ;
 
-    UserAccountStateEnum(Integer value, String name, String info) {
+    UserAccountStateEnum(Short value, String name, String info) {
         this.value = value;
         this.name = name;
         this.info = info;
     }
 
-    private Integer value ;
+    private Short value ;
     private String name ;
     private String info ;
 
 
-    public static String doGetEnumInfoByValue(Integer value){
+    public static String doGetEnumInfoByValue(Short value){
         if(value == null){
             return null ;
         }
@@ -33,11 +33,11 @@ public enum UserAccountStateEnum {
 
 
 
-    public Integer getValue() {
+    public Short getValue() {
         return value;
     }
 
-    public void setValue(Integer value) {
+    public void setValue(Short value) {
         this.value = value;
     }
