+++ egg-persistence/src/main/java/com/egg/manager/persistence/entity/user/UserTenant.java
@@ -29,10 +29,10 @@ public class UserTenant extends Model<UserTenant> {
     private  String defineTenantId ;
     private Integer type;
     @TableField(value = "is_manager")
-    private Integer isManager;
+    private Short isManager;
 
     private String remark;
-    private Integer state ;
+    private Short state ;
     @TableField("create_time")
     private Date createTime ;
     @TableField("update_time")
