+++ egg-persistence/src/main/java/com/egg/manager/persistence/mapper/define/DefineMenuMapper.java
@@ -36,14 +36,14 @@ public interface DefineMenuMapper extends BaseMapper<DefineMenu> {
      * @param roleId
      * @return
      */
-    List<DefineMenu> findAllMenuByRoleId(@Param("roleId")String roleId,@Param("stateVal")Integer state) ;
+    List<DefineMenu> findAllMenuByRoleId(@Param("roleId")String roleId,@Param("stateVal")Short state) ;
 
     /**
      * 取得角色拥有的所有[菜单]集合(忽略 有子节点的菜单节点)
      * @param roleId
      * @return
      */
-    List<DefineMenu> findAllMenuByRoleIdFilterParentNode(@Param("roleId")String roleId,@Param("stateVal")Integer state) ;
+    List<DefineMenu> findAllMenuByRoleIdFilterParentNode(@Param("roleId")String roleId,@Param("stateVal")Short state) ;
 
 
     /**
