+++ egg-persistence/src/main/java/com/egg/manager/persistence/vo/user/UserRoleVo.java
@@ -21,7 +21,7 @@ public class UserRoleVo {
     private Integer type;
 
     private String remark;
-    private Integer state;
+    private Short state;
     private Date createTime;
     private Date updateTime;
     private String createUserId ;
